validation & bundle creation in separate service task, validation ToDos
