+++ mii-dsf-process-projectathon-data-transfer/src/main/java/de/medizininformatik_initiative/processes/projectathon/data_transfer/service/ValidateDataDic.java
@@ -1,17 +1,6 @@
 package de.medizininformatik_initiative.processes.projectathon.data_transfer.service;
 
-import static de.medizininformatik_initiative.processes.projectathon.data_transfer.ConstantsDataTransfer.BPMN_EXECUTION_VARIABLE_BINARY;
-import static de.medizininformatik_initiative.processes.projectathon.data_transfer.ConstantsDataTransfer.BPMN_EXECUTION_VARIABLE_DATA_SET;
-import static de.medizininformatik_initiative.processes.projectathon.data_transfer.ConstantsDataTransfer.BPMN_EXECUTION_VARIABLE_DOCUMENT_REFERENCE;
-import static de.medizininformatik_initiative.processes.projectathon.data_transfer.ConstantsDataTransfer.BPMN_EXECUTION_VARIABLE_PROJECT_IDENTIFIER;
-import static de.medizininformatik_initiative.processes.projectathon.data_transfer.ConstantsDataTransfer.NAMINGSYSTEM_MII_PROJECT_IDENTIFIER;
-import static org.highmed.dsf.bpe.ConstantsBase.NAMINGSYSTEM_HIGHMED_ORGANIZATION_IDENTIFIER;
-import static org.hl7.fhir.r4.model.Bundle.BundleType.TRANSACTION;
-import static org.hl7.fhir.r4.model.DocumentReference.ReferredDocumentStatus.FINAL;
-import static org.hl7.fhir.r4.model.Enumerations.DocumentReferenceStatus.CURRENT;
-
 import java.util.Objects;
-import java.util.UUID;
 
 import org.camunda.bpm.engine.delegate.DelegateExecution;
 import org.highmed.dsf.bpe.delegate.AbstractServiceDelegate;
@@ -19,21 +8,10 @@ import org.highmed.dsf.fhir.authorization.read.ReadAccessHelper;
 import org.highmed.dsf.fhir.client.FhirWebserviceClientProvider;
 import org.highmed.dsf.fhir.organization.OrganizationProvider;
 import org.highmed.dsf.fhir.task.TaskHelper;
-import org.highmed.dsf.fhir.variables.FhirResourceValues;
-import org.hl7.fhir.r4.model.Binary;
-import org.hl7.fhir.r4.model.Bundle;
-import org.hl7.fhir.r4.model.DocumentReference;
-import org.hl7.fhir.r4.model.ResourceType;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.InitializingBean;
-
-import ca.uhn.fhir.context.FhirContext;
 
 public class ValidateDataDic extends AbstractServiceDelegate implements InitializingBean
 {
-	private static final Logger logger = LoggerFactory.getLogger(ValidateDataDic.class);
-
 	private final OrganizationProvider organizationProvider;
 
 	public ValidateDataDic(FhirWebserviceClientProvider clientProvider, TaskHelper taskHelper,
@@ -55,40 +33,19 @@ public class ValidateDataDic extends AbstractServiceDelegate implements Initiali
 	@Override
 	protected void doExecute(DelegateExecution execution)
 	{
-		String projectIdentifier = (String) execution.getVariable(BPMN_EXECUTION_VARIABLE_PROJECT_IDENTIFIER);
-		DocumentReference documentReference = (DocumentReference) execution
-				.getVariable(BPMN_EXECUTION_VARIABLE_DOCUMENT_REFERENCE);
-		Binary binary = (Binary) execution.getVariable(BPMN_EXECUTION_VARIABLE_BINARY);
+		// String projectIdentifier = (String) execution.getVariable(BPMN_EXECUTION_VARIABLE_PROJECT_IDENTIFIER);
+		// DocumentReference documentReference = (DocumentReference) execution
+		// .getVariable(BPMN_EXECUTION_VARIABLE_DOCUMENT_REFERENCE);
+		// Binary binary = (Binary) execution.getVariable(BPMN_EXECUTION_VARIABLE_BINARY);
+		//
+		// binary.getContentType();
+		// binary.getData();
 
-		Bundle bundle = createTransactionBundle(projectIdentifier, documentReference, binary);
-		logger.debug("Created Bundle: {}", FhirContext.forR4().newXmlParser().encodeResourceToString(bundle));
-
-		execution.setVariable(BPMN_EXECUTION_VARIABLE_DATA_SET, FhirResourceValues.create(bundle));
-	}
-
-	private Bundle createTransactionBundle(String projectIdentifier, DocumentReference documentReference, Binary binary)
-	{
-		Binary binaryToTransmit = new Binary().setContentType(binary.getContentType());
-		binaryToTransmit.setContent(binary.getContent());
-		binaryToTransmit.setId(UUID.randomUUID().toString());
-
-		DocumentReference documentReferenceToTransmit = new DocumentReference().setStatus(CURRENT).setDocStatus(FINAL);
-		documentReferenceToTransmit.getMasterIdentifier().setSystem(NAMINGSYSTEM_MII_PROJECT_IDENTIFIER)
-				.setValue(projectIdentifier);
-		documentReferenceToTransmit.addAuthor().setType(ResourceType.Organization.name()).getIdentifier()
-				.setSystem(NAMINGSYSTEM_HIGHMED_ORGANIZATION_IDENTIFIER)
-				.setValue(organizationProvider.getLocalIdentifierValue());
-		documentReferenceToTransmit.setDate(documentReference.getDate());
-		documentReferenceToTransmit.addContent().getAttachment().setContentType(binary.getContentType())
-				.setUrl("urn:uuid:" + binaryToTransmit.getId());
-
-		Bundle bundle = new Bundle().setType(TRANSACTION);
-		bundle.addEntry().setResource(documentReferenceToTransmit)
-				.setFullUrl("urn:uuid:" + documentReferenceToTransmit.getId()).getRequest()
-				.setMethod(Bundle.HTTPVerb.POST).setUrl(ResourceType.DocumentReference.name());
-		bundle.addEntry().setResource(binaryToTransmit).setFullUrl("urn:uuid:" + binaryToTransmit.getId()).getRequest()
-				.setMethod(Bundle.HTTPVerb.POST).setUrl(ResourceType.Binary.name());
-
-		return bundle;
+		/*
+		 * TODO validate declared mime-type = text/csv, if possible validate actual mime-type of data for example using
+		 * Apache Tika -> https://tika.apache.org/1.21/detection.html
+		 * 
+		 * ...possible transfer of malicious binary data
+		 */
 	}
 }
